fxemu: compute SuperFX speedPerLine with integer math (determinism)
The SuperFX (GSU) per-scanline instruction budget was derived through
float32 arithmetic in S9xResetSuperFX:

    speedPerLine = (uint32_t)(SuperFXSpeedPerLine * ((1.0f/fps)/V_Max));

The SNES has no FPU, so nothing here models hardware; the float math is
merely an approximation whose truncated result is not guaranteed to be
bit-identical across x87/SSE/ARM or compiler flags. Since speedPerLine
directly controls how many GSU instructions execute per line, any drift
diverges emulation and breaks netplay/savestate/runahead determinism.

Reformulate as exact integer division. SuperFXSpeedPerLine is now stored
as uint32_t cycles/second (the 0.417 duty factor and MHz scaling fold to
exact integers: default 0.417*10.5e6 == 4378500; overclock option is
integer MHz only, 0.417*1.5e6 == 625500 cycles/s per MHz). The reset path
becomes:

    speedPerLine = SuperFXSpeedPerLine / (fps * V_Max);

with fps the fixed 50/60 the core already used and V_Max (262/312) set by
S9xResetCPU beforehand, so the divisor is never zero. Integer division
truncates toward zero, matching the prior cast.

Verified the integer result is identical to the old float truncation for
every reachable configuration (default + all 12 overclock MHz x NTSC/PAL);
none lands near an integer boundary, so this is behavior-preserving today
while removing the nondeterminism hazard. objdump of S9xResetSuperFX
confirms no FP arithmetic or int<->float conversions remain; the per-line
budget is now an integer div. Full LTO build links clean.

+++ src/snes9x.h
@@ -430,7 +430,7 @@ struct SSettings
 	   Toons rainbow rings). Stable matches HD-no-BL's Y-axis sampling
 	   and only smooths X. */
 	int32_t		Mode7HiresBilinear;
-   float    SuperFXSpeedPerLine;
+   uint32_t SuperFXSpeedPerLine; /* GSU clock cycles per second (already scaled by the 0.417 duty factor) */
    uint8_t	FastSavestates;
    uint8_t	HardDisableAudio;
 };
